Finished 44

+++ utils/proj_eul_math/general.pxd
@@ -3,7 +3,7 @@ import cython
 @cython.locals(total_sum=cython.ulong, current_sum=cython.ulong)
 cpdef unsigned long int get_sum_divisors(unsigned int num=*)
 @cython.locals(x=cython.uint, seen=set)
-cpdef unsigned int is_square(unsigned int n)
+cpdef unsigned int is_square(unsigned long long n)
 @cython.locals(total_divisors=cython.uint)
 cpdef unsigned int get_num_divisors(unsigned int num=*)
 @cython.locals(temp=cython.uint)
